requirements.txt
aliyun-python-sdk-core-v3==2.13.11
astroid==2.3.2
bleach==3.1.0
certifi==2019.9.11
chardet==3.0.4
colorama==0.4.1
Django==2.2.6
django-debug-toolbar==2.1
django-haystack==2.8.1
django-redis==4.10.0
django-shortuuidfield==0.1.3
djangorestframework==3.11.0
idna==2.8
isort==4.3.21
jieba==0.39
jmespath==0.9.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
mysqlclient==1.4.5
Pillow==6.2.1
pylint==2.4.3
pylint-django==2.0.12
pylint-plugin-utils==0.6
python-memcached==1.59
pytz==2019.3
qiniu==7.2.4
redis==3.3.11
requests==2.22.0
shortuuid==0.5.0
six==1.12.0
sqlparse==0.3.0
typed-ast==1.4.0
urllib3==1.25.6
webencodings==0.5.1
Whoosh==2.7.4
wrapt==1.11.1
